+++ lib/config/flows/type.ex
@@ -1,18 +1,30 @@
 defmodule FormFlow.Config.Flows.Type do
+  @moduledoc """
+  Flow type definition
+  """
+
   defstruct [:id, :module, :name, :description, :properties]
 
+  @type t :: %__MODULE__{
+          id: String.t(),
+          module: module(),
+          name: String.t(),
+          description: String.t() | nil,
+          properties: keyword() | map() | nil
+        }
+
+  @doc "Form progress component, return `nil` to not show"
   @callback form_progress_component(map()) :: any()
 
   defmacro __using__(_opts) do
     quote do
       @behaviour FormFlow.Config.Flows.Type
 
-      @doc "Form progress component, return `nil` to not show"
       def form_progress_component(assigns) do
-        FormFlow.Web.Instances.Components.Flows.Types.Default.form_progress_component(assigns)
+        FormFlow.Web.Components.Flows.Types.Default.form_progress_component(assigns)
       end
 
-      def overridable [form_progress_component: 1]
+      defoverridable form_progress_component: 1
     end
   end
 end