+++ lib/web/components/flows/types_wizard_any_order.ex
@@ -1,3 +1,8 @@
-defmodule FormFlow.Web.Instances.Components.Flows.Types.WizardAnyOrder do
+defmodule FormFlow.Web.Components.Flows.Types.WizardAnyOrder do
+  @moduledoc """
+  Flow type `"wizard_any_order"`: a "forms" flow whose forms can be completed
+  in any order.
+  """
+
   use FormFlow.Config.Flows.Type
 end