+++ lib/web/components/flows/types_wizard_in_order.ex
@@ -1,3 +1,8 @@
-defmodule FormFlow.Web.Instances.Components.Flows.Types.WizardInOrder do
+defmodule FormFlow.Web.Components.Flows.Types.WizardInOrder do
+  @moduledoc """
+  Flow type `"wizard_in_order"`: a "forms" flow whose forms are completed
+  front to back.
+  """
+
   use FormFlow.Config.Flows.Type
 end