+++ lib/web/live/instances/forms/show.ex
@@ -31,8 +31,6 @@ defmodule FormFlow.Web.Instances.Forms.Show do
       socket
       |> assign(assigns)
       |> assign_new(:base, fn -> "" end)
-      |> assign_new(:config, fn -> nil end)
-      |> assign_new(:config_data, fn -> %{} end)
       |> assign_new(:error, fn -> nil end)
 
     {:ok, load(socket)}